OMAP: DSS2: GFX FIFO UNDERFLOW issue fixed
In case of 720P with 90/270 degree rotation, the system reports
GFX_FIFO_UNDERFLOW error which usually happens if DSS DMA is not able to fill
the FIFO as per requirement.

In TRM (section 11.2.6.1.3), where is has been clearly mentioned that,

"To improve the performance on 90 degree rotation, split the data access on
write side and not read side."

That means, read should always happen on 0 degree and write should go to
respective rotation view.

Change-Id: I04f4c7cb4d6ae89d4762f793dd7b60db8893a1a6
Signed-off-by: Sriramakrishnan A G <srk@ti.com>